Clean up a bunch of the code
The goal here is to avoid formatting issues with PRs in the future such
as parameter placement, spacing, etc.

+++ src/TaglibSharp.Tests/Performance/PerformanceTest.cs
@@ -1,5 +1,5 @@
-using System;
 using NUnit.Framework;
+using System;
 using TagLib;
 
 namespace TaglibSharp.Tests.Performance
@@ -8,46 +8,46 @@ namespace TaglibSharp.Tests.Performance
 	public class PerformanceTest
 	{
 		[Test]
-		public void CreateM4a()
+		public void CreateM4a ()
 		{
 			try {
 				double total_time = 0.0;
 				int iterations = 1000;
-				using(new CodeTimer("Combined")) {
-					for(int i = 0; i < iterations; i++) {
-						var timer = new CodeTimer();
-						using(timer) {
-							File.Create(TestPath.Samples + "sample.m4a");
+				using (new CodeTimer ("Combined")) {
+					for (int i = 0; i < iterations; i++) {
+						var timer = new CodeTimer ();
+						using (timer) {
+							File.Create (TestPath.Samples + "sample.m4a");
 						}
 						total_time += timer.ElapsedTime.TotalSeconds;
 					}
 				}
-				Console.WriteLine("Average time: {0}", total_time / (double)iterations);
-			} catch(Exception e) {
-				Console.WriteLine(e);
+				Console.WriteLine ("Average time: {0}", total_time / iterations);
+			} catch (Exception e) {
+				Console.WriteLine (e);
 				Assert.False (false);
 			}
 			Assert.True (true);
 		}
 
 		[Test]
-		public void CreateOgg()
+		public void CreateOgg ()
 		{
 			try {
 				double total_time = 0.0;
 				int iterations = 1000;
-				using(new CodeTimer("Combined")) {
-					for(int i = 0; i < iterations; i++) {
-						var timer = new CodeTimer();
-						using(timer) {
-							File.Create(TestPath.Samples + "sample.ogg");
+				using (new CodeTimer ("Combined")) {
+					for (int i = 0; i < iterations; i++) {
+						var timer = new CodeTimer ();
+						using (timer) {
+							File.Create (TestPath.Samples + "sample.ogg");
 						}
 						total_time += timer.ElapsedTime.TotalSeconds;
 					}
 				}
-				Console.WriteLine("Average time: {0}", total_time / (double)iterations);
-			} catch(Exception e) {
-				Console.WriteLine(e);
+				Console.WriteLine ("Average time: {0}", total_time / iterations);
+			} catch (Exception e) {
+				Console.WriteLine (e);
 				Assert.False (false);
 			}
 			Assert.True (true);
